Move MD mode buttons to left also
